+++ internal/errors/errors.go
@@ -58,12 +58,12 @@ func Is(err, target error) bool {
 // New is a proxy to errors.New
 // This need to single import errors
 func New(text string) error {
-	return Errorf(0, "%w", errors.New(text))
+	return ErrorfSkip(1, "%w", errors.New(text))
 }
 
 // NewWithIssues returns error which contains child issues
 func NewWithIssues(text string, issues ...error) error {
-	return Errorf(1, "%w", &errorWithIssues{
+	return ErrorfSkip(1, "%w", &errorWithIssues{
 		reason: text,
 		issues: issues,
 	})
@@ -89,7 +89,12 @@ func (e *errorWithIssues) Error() string {
 }
 
 // Errorf is alias to fmt.Errorf() with prepend file:line prefix
-func Errorf(depth int, format string, args ...interface{}) error {
+func Errorf(format string, args ...interface{}) error {
+	return ErrorfSkip(1, format, args...)
+}
+
+// ErrorfSkip is alias to fmt.Errorf() with prepend file:line prefix
+func ErrorfSkip(depth int, format string, args ...interface{}) error {
 	function, file, line, _ := runtime.Caller(depth + 1)
 	return fmt.Errorf(runtime.FuncForPC(function).Name()+" ("+fileName(file)+":"+strconv.Itoa(line)+") "+format, args...)
 }