+++ internal/meta/meta.go
@@ -121,7 +121,7 @@ func (m *meta) meta(ctx context.Context) (_ metadata.MD, err error) {
 	token, err = m.credentials.Token(ctx)
 	if err != nil {
 		if stringer, ok := m.credentials.(fmt.Stringer); ok {
-			return nil, errors.Errorf(0, "%s: %w", stringer.String(), err)
+			return nil, errors.Errorf("%s: %w", stringer.String(), err)
 		}
 		return nil, err
 	}