+++ sugar/sugar.go
@@ -34,12 +34,12 @@ func MakeRecursive(ctx context.Context, db ydb.Connection, pathToCreate string) 
 		if errors.As(err, &opErr) && opErr.Reason == errors.StatusSchemeError {
 			err = db.Scheme().MakeDirectory(ctx, sub)
 			if err != nil {
-				return errors.Errorf(0, "%w", err)
+				return errors.Errorf("%w", err)
 			}
 			info, err = db.Scheme().DescribePath(ctx, sub)
 		}
 		if err != nil {
-			return errors.Errorf(0, "%w", err)
+			return errors.Errorf("%w", err)
 		}
 		switch info.Type {
 		case
@@ -47,7 +47,7 @@ func MakeRecursive(ctx context.Context, db ydb.Connection, pathToCreate string) 
 			scheme.EntryDirectory:
 			// OK
 		default:
-			return errors.Errorf(0, "entry %q exists but it is a %s",
+			return errors.Errorf("entry %q exists but it is a %s",
 				sub, info.Type,
 			)
 		}
@@ -76,7 +76,7 @@ func RemoveRecursive(ctx context.Context, db ydb.Connection, pathToRemove string
 			return nil
 		}
 		if err != nil {
-			return errors.Errorf(0, "%w", err)
+			return errors.Errorf("%w", err)
 		}
 
 		for _, child := range dir.Children {
@@ -87,13 +87,13 @@ func RemoveRecursive(ctx context.Context, db ydb.Connection, pathToRemove string
 			switch child.Type {
 			case scheme.EntryDirectory:
 				if err = list(i+1, pt); err != nil {
-					return errors.Errorf(0, "%w", err)
+					return errors.Errorf("%w", err)
 				}
 				err = retry.Retry(ctx, func(ctx context.Context) (err error) {
 					return db.Scheme().RemoveDirectory(ctx, pt)
 				}, retry.WithIdempotent())
 				if err != nil {
-					return errors.Errorf(0, "%w", err)
+					return errors.Errorf("%w", err)
 				}
 
 			case scheme.EntryTable:
@@ -101,7 +101,7 @@ func RemoveRecursive(ctx context.Context, db ydb.Connection, pathToRemove string
 					return session.DropTable(ctx, pt)
 				})
 				if err != nil {
-					return errors.Errorf(0, "%w", err)
+					return errors.Errorf("%w", err)
 				}
 
 			default: